Diffusers.Schedulers.DDPM — Type

Denoising Diffusion Probabilistic Models (DDPM) scheduler.

References

[2006.11239] Denoising Diffusion Probabilistic Models

Diffusers.Schedulers.Scheduler — Type

Abstract type for schedulers.

Diffusers.Schedulers.add_noise — Method

Add noise to clean data using the forward diffusion process.

Input

scheduler::Scheduler: scheduler to use

x₀::AbstractArray: clean data to add noise to

ϵ::AbstractArray: noise to add to clean data

t::AbstractArray: timesteps used to weight the noise

Output

xₜ::AbstractArray: noisy data at the given timesteps

Diffusers.Schedulers.step — Method

Remove noise from model output using the backward diffusion process.

Input

scheduler::Scheduler: scheduler to use

xₜ::AbstractArray: sample to be denoised

ϵᵧ::AbstractArray: predicted noise to remove

t::AbstractArray: timestep t of xₜ

Output

xₜ₋₁::AbstractArray: denoised sample at t=t-1

x̂₀::AbstractArray: denoised sample at t=0